little tweaks mostly.
Big thing to note: Using BOM in UTF-8 is apparently not recommended, and
is really only intended for backwards compatibility, to mark the start
of a stream.

When I first starting extracting TODOs (several projects ago), I had
assumed that the default constructor of the encoding was equivalent to
not using a BOM. Using the default .ctor meant that Notepad++ couldn't
guess the encoding. That was the only reason I set the encoding to use
the BOM in the first place. Turns out, that default .ctor really is
neither, but setting it to false explicitly works fine.